refactoring and icons
