+++ frontend/src/context/UserContext.jsx
@@ -1,40 +1,3 @@
-import React, { createContext, useState, useContext, useEffect } from 'react';
+import { createContext } from 'react';
 
-const UserContext = createContext();
-
-export const useUser = () => {
-  return useContext(UserContext);
-};
-
-export const UserProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      const storedUser = localStorage.getItem('user'); 
-      if (storedUser) {
-        setUser(JSON.parse(storedUser));
-      }
-      setLoading(false);
-    };
-
-    fetchUser();
-  }, []);
-
-  const login = userData => {
-    setUser(userData); 
-    localStorage.setItem('user', JSON.stringify(userData));
-  };
-
-  const logout = () => {
-    setUser(null);
-    localStorage.removeItem('user');
-  };
-
-  return (
-    <UserContext.Provider value={{ user, loading, setUser, login, logout }}>
-      {children}
-    </UserContext.Provider>
-  );
-};
+export const UserContext = createContext();
