+++ frontend/src/context/UserProvider.js
@@ -1,0 +1,35 @@
+import React, { useState, useEffect } from 'react';
+import { UserContext } from './UserContext';
+
+export const UserProvider = ({ children }) => {
+  const [user, setUser] = useState(null);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchUser = async () => {
+      const storedUser = localStorage.getItem('user');
+      if (storedUser) {
+        setUser(JSON.parse(storedUser));
+      }
+      setLoading(false);
+    };
+
+    fetchUser();
+  }, []);
+
+  const login = userData => {
+    setUser(userData);
+    localStorage.setItem('user', JSON.stringify(userData));
+  };
+
+  const logout = () => {
+    setUser(null);
+    localStorage.removeItem('user');
+  };
+
+  return (
+    <UserContext.Provider value={{ user, loading, login, logout, setUser }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
