--- frontend/src/data/categories.js
@@ -1,47 +1,0 @@
-const categories = [
-  {
-    name: 'Beef',
-    image: '/image/categories/beef-1x.png',
-  },
-  {
-    name: 'Breakfast',
-    image: '/image/categories/breakfast-1x.png',
-  },
-  {
-    name: 'Desserts',
-    image: '/image/categories/desserts-1x.png',
-  },
-  {
-    name: 'Lamb',
-    image: '/image/categories/lamb-1x.png',
-  },
-  {
-    name: 'Goat',
-    image: '/image/categories/goat-1x.png',
-  },
-  {
-    name: 'Miscellaneous',
-    image: '/image/categories/miscellaneous-1x.png',
-  },
-  {
-    name: 'Pasta',
-    image: '/image/categories/pasta-1x.png',
-  },
-  {
-    name: 'Pork',
-    image: '/image/categories/pork-1x.png',
-  },
-  {
-    name: 'Seafood',
-    image: '/image/categories/seafood-1x.png',
-  },
-  {
-    name: 'Side',
-    image: '/image/categories/side-1x.png',
-  },
-  {
-    name: 'Starter',
-    image: '/image/categories/starter-1x.png',
-  },
-];
-export default categories;
