+++ frontend/src/hooks/useUser.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { UserContext } from '../contexts/UserContext';
+
+export const useUser = () => {
+  return useContext(UserContext);
+};
